[dynamodb] Move store methods
