+++ sources/dynamodb/snapshot.go
@@ -7,11 +7,39 @@ import (
 
 	"github.com/aws/aws-sdk-go/service/dynamodb"
 
+	"github.com/artie-labs/reader/config"
 	"github.com/artie-labs/reader/lib"
 	"github.com/artie-labs/reader/lib/dynamo"
 	"github.com/artie-labs/reader/lib/kafkalib"
 	"github.com/artie-labs/reader/lib/logger"
+	"github.com/artie-labs/reader/lib/s3lib"
 )
+
+type SnapshotStore struct {
+	tableName string
+	streamArn string
+	cfg       *config.DynamoDB
+
+	s3Client       *s3lib.S3Client
+	dynamoDBClient *dynamodb.DynamoDB
+}
+
+func (s *SnapshotStore) Close() error {
+	return nil
+}
+
+func (s *SnapshotStore) Run(ctx context.Context, writer kafkalib.BatchWriter) error {
+	if err := s.scanFilesOverBucket(); err != nil {
+		return fmt.Errorf("scanning files over bucket failed: %w", err)
+	}
+
+	if err := s.streamAndPublish(ctx, writer); err != nil {
+		return fmt.Errorf("stream and publish failed: %w", err)
+	}
+
+	slog.Info("Finished snapshotting all the files")
+	return nil
+}
 
 func (s *SnapshotStore) scanFilesOverBucket() error {
 	if len(s.cfg.SnapshotSettings.SpecifiedFiles) > 0 {
